LPS-105162 move 1_1_2 to 1_1_3
