+++ modules/dxp/apps/saml/saml-persistence-service/src/main/java/com/liferay/saml/persistence/internal/upgrade/v1_1_3/util/SamlSpIdpConnectionTable.java
@@ -12,7 +12,7 @@
  *
  */
 
-package com.liferay.saml.persistence.internal.upgrade.v1_1_2.util;
+package com.liferay.saml.persistence.internal.upgrade.v1_1_3.util;
 
 import java.sql.Types;
